chore(repo): split slow e2e tests
