+++ e2e/angular/src/cypress-component-tests-app.test.ts
@@ -1,0 +1,32 @@
+import {
+  runCLI,
+  runE2ETests,
+} from '@nx/e2e-utils';
+import {
+  setupCypressComponentTests,
+  cleanupCypressComponentTests,
+  CypressComponentTestsSetup,
+} from './cypress-component-tests-setup';
+
+describe('Angular Cypress Component Tests - App', () => {
+  let setup: CypressComponentTestsSetup;
+
+  beforeAll(async () => {
+    setup = setupCypressComponentTests();
+  });
+
+  afterAll(() => cleanupCypressComponentTests());
+
+  it('should test app', () => {
+    const { appName } = setup;
+    
+    runCLI(
+      `generate @nx/angular:cypress-component-configuration --project=${appName} --generate-tests --no-interactive`
+    );
+    if (runE2ETests('cypress')) {
+      expect(runCLI(`component-test ${appName}`)).toContain(
+        'All specs passed!'
+      );
+    }
+  }, 300_000);
+});
