+++ e2e/angular/src/cypress-component-tests-buildable.test.ts
@@ -1,0 +1,67 @@
+import {
+  checkFilesDoNotExist,
+  runCLI,
+  runE2ETests,
+  updateFile,
+} from '@nx/e2e-utils';
+import {
+  setupCypressComponentTests,
+  cleanupCypressComponentTests,
+  updateTestToAssertTailwindIsNotApplied,
+  CypressComponentTestsSetup,
+} from './cypress-component-tests-setup';
+
+describe('Angular Cypress Component Tests - Buildable Lib', () => {
+  let setup: CypressComponentTestsSetup;
+
+  beforeAll(async () => {
+    setup = setupCypressComponentTests();
+  });
+
+  afterAll(() => cleanupCypressComponentTests());
+
+  it('should test buildable lib not being used in app', () => {
+    const { appName, buildableLibName } = setup;
+    
+    expect(() => {
+      // should error since no edge in graph between lib and app
+      runCLI(
+        `generate @nx/angular:cypress-component-configuration --project=${buildableLibName} --generate-tests --no-interactive`
+      );
+    }).toThrow();
+
+    updateTestToAssertTailwindIsNotApplied(buildableLibName);
+
+    runCLI(
+      `generate @nx/angular:cypress-component-configuration --project=${buildableLibName} --generate-tests --build-target=${appName}:build --no-interactive`
+    );
+    if (runE2ETests('cypress')) {
+      expect(runCLI(`component-test ${buildableLibName}`)).toContain(
+        'All specs passed!'
+      );
+    }
+    // add tailwind
+    runCLI(`generate @nx/angular:setup-tailwind --project=${buildableLibName}`);
+    updateFile(
+      `${buildableLibName}/src/lib/input/input.component.cy.ts`,
+      (content) => {
+        // text-green-500 should now apply
+        return content.replace('rgb(0, 0, 0)', 'rgb(34, 197, 94)');
+      }
+    );
+    updateFile(
+      `${buildableLibName}/src/lib/input-standalone/input-standalone.cy.ts`,
+      (content) => {
+        // text-green-500 should now apply
+        return content.replace('rgb(0, 0, 0)', 'rgb(34, 197, 94)');
+      }
+    );
+
+    if (runE2ETests('cypress')) {
+      expect(runCLI(`component-test ${buildableLibName}`)).toContain(
+        'All specs passed!'
+      );
+      checkFilesDoNotExist(`tmp${buildableLibName}/ct-styles.css`);
+    }
+  }, 300_000);
+});
