+++ e2e/angular/src/cypress-component-tests-lib.test.ts
@@ -1,0 +1,32 @@
+import {
+  runCLI,
+  runE2ETests,
+} from '@nx/e2e-utils';
+import {
+  setupCypressComponentTests,
+  cleanupCypressComponentTests,
+  CypressComponentTestsSetup,
+} from './cypress-component-tests-setup';
+
+describe('Angular Cypress Component Tests - Lib', () => {
+  let setup: CypressComponentTestsSetup;
+
+  beforeAll(async () => {
+    setup = setupCypressComponentTests();
+  });
+
+  afterAll(() => cleanupCypressComponentTests());
+
+  it('should successfully component test lib being used in app', () => {
+    const { usedInAppLibName } = setup;
+    
+    runCLI(
+      `generate @nx/angular:cypress-component-configuration --project=${usedInAppLibName} --generate-tests --no-interactive`
+    );
+    if (runE2ETests('cypress')) {
+      expect(runCLI(`component-test ${usedInAppLibName}`)).toContain(
+        'All specs passed!'
+      );
+    }
+  }, 300_000);
+});
