+++ e2e/esbuild/src/esbuild-basic.test.ts
@@ -1,14 +1,9 @@
 import {
   checkFilesDoNotExist,
   checkFilesExist,
-  cleanupProject,
-  createFile,
   detectPackageManager,
-  newProject,
-  packageInstall,
   packageManagerLockFile,
   readFile,
-  readJson,
   runCLI,
   runCommand,
   tmpProjPath,
@@ -17,13 +12,14 @@ import {
   updateJson,
 } from '@nx/e2e-utils';
 import { join } from 'path';
+import { setupEsbuildTest, cleanupEsbuildTest } from './esbuild-setup';
 
-describe('EsBuild Plugin', () => {
+describe('EsBuild Plugin - Basic', () => {
   let proj: string;
 
-  beforeEach(() => (proj = newProject()));
+  beforeEach(() => (proj = setupEsbuildTest()));
 
-  afterEach(() => cleanupProject());
+  afterEach(() => cleanupEsbuildTest());
 
   it('should setup and build projects using build', async () => {
     const myPkg = uniq('my-pkg');
@@ -40,9 +36,9 @@ describe('EsBuild Plugin', () => {
 
     // Copy package.json as asset rather than generate with Nx-detected fields.
     runCLI(`build ${myPkg} --generatePackageJson=false`);
-    const packageJson = readJson(`libs/${myPkg}/package.json`);
+    const packageJson = readFile(`libs/${myPkg}/package.json`);
     // This is the file that is generated by lib generator (no deps, no main, etc.).
-    expect(packageJson).toEqual({
+    expect(JSON.parse(packageJson)).toEqual({
       name: `@proj/${myPkg}`,
       version: '0.0.1',
       private: true,
@@ -121,61 +117,6 @@ describe('EsBuild Plugin', () => {
     //   )
     // ).resolves.not.toThrow();
     // watchProcess.kill();
-  }, 300_000);
-
-  it('should support bundling everything or only workspace libs', async () => {
-    packageInstall('rambda', undefined, '~7.3.0', 'prod');
-    packageInstall('lodash', undefined, '~4.14.0', 'prod');
-    const parentLib = uniq('parent-lib');
-    const childLib = uniq('child-lib');
-    runCLI(
-      `generate @nx/js:lib ${parentLib} --directory=libs/${parentLib} --bundler=esbuild`
-    );
-    runCLI(
-      `generate @nx/js:lib ${childLib} --directory=libs/${childLib} --bundler=none`
-    );
-    updateFile(
-      `libs/${parentLib}/src/index.ts`,
-      `
-        // @ts-ignore
-        import _ from 'lodash';
-        import { greet } from '@${proj}/${childLib}';
-
-        console.log(_.upperFirst('hello world'));
-        console.log(greet());
-      `
-    );
-    updateFile(
-      `libs/${childLib}/src/index.ts`,
-      `
-        import { always } from 'rambda';
-        export const greet = always('Hello from child lib');
-      `
-    );
-
-    // Bundle child lib and third-party packages
-    runCLI(`build ${parentLib}`);
-
-    expect(
-      readJson(`dist/libs/${parentLib}/package.json`).dependencies?.['dayjs']
-    ).not.toBeDefined();
-    let runResult = runCommand(`node dist/libs/${parentLib}/index.cjs`);
-    expect(runResult).toMatch(/Hello world/);
-    expect(runResult).toMatch(/Hello from child lib/);
-
-    // Bundle only child lib
-    runCLI(`build ${parentLib} --third-party=false`);
-
-    expect(
-      readJson(`dist/libs/${parentLib}/package.json`).dependencies
-    ).toEqual({
-      // Don't care about the versions, just that they exist
-      rambda: expect.any(String),
-      lodash: expect.any(String),
-    });
-    runResult = runCommand(`node dist/libs/${parentLib}/index.cjs`);
-    expect(runResult).toMatch(/Hello world/);
-    expect(runResult).toMatch(/Hello from child lib/);
   }, 300_000);
 
   it('should support non-bundle builds', () => {
@@ -277,40 +218,4 @@ describe('EsBuild Plugin', () => {
     expect(output).not.toMatch(/secret/);
     expect(output).toMatch(/foobar/);
   });
-
-  it('should support declaration builds', () => {
-    const declarationPkg = uniq('declaration-pkg');
-    runCLI(
-      `generate @nx/js:lib ${declarationPkg} --directory=libs/${declarationPkg} --bundler=esbuild`
-    );
-    createFile(
-      `libs/${declarationPkg}/src/lib/testDir/sub.ts`,
-      `
-        export function sub(): string {
-          return 'sub';
-        }
-      `
-    );
-    updateFile(
-      `libs/${declarationPkg}/src/lib/${declarationPkg}.ts`,
-      `
-        import { sub } from './testDir/sub';
-        
-        console.log('${declarationPkg}-' + sub());
-      `
-    );
-
-    runCLI(`build ${declarationPkg} --declaration=true`);
-
-    checkFilesExist(
-      `dist/libs/${declarationPkg}/index.cjs`,
-      `dist/libs/${declarationPkg}/src/index.d.ts`,
-      `dist/libs/${declarationPkg}/src/lib/${declarationPkg}.d.ts`,
-      `dist/libs/${declarationPkg}/src/lib/testDir/sub.d.ts`
-    );
-
-    expect(runCommand(`node dist/libs/${declarationPkg}`)).toMatch(
-      new RegExp(`${declarationPkg}-sub`)
-    );
-  }, 300_000);
 });
