+++ e2e/esbuild/src/esbuild-setup.ts
@@ -1,0 +1,9 @@
+import { cleanupProject, newProject } from '@nx/e2e-utils';
+
+export function setupEsbuildTest(): string {
+  return newProject();
+}
+
+export function cleanupEsbuildTest(): void {
+  cleanupProject();
+}
