+++ e2e/react/src/module-federation/core-rspack-name-and-root.test.ts
@@ -1,0 +1,81 @@
+import {
+  checkFilesExist,
+  killProcessAndPorts,
+  uniq,
+  runCommandUntil,
+} from '@nx/e2e-utils';
+import { readPort, runCLI } from './utils';
+import {
+  setupCoreRspackTest,
+  cleanupCoreRspackTest,
+} from './core-rspack-setup';
+
+describe('React Rspack Module Federation - Name and Root Format', () => {
+  beforeEach(() => {
+    setupCoreRspackTest();
+  });
+
+  afterEach(() => cleanupCoreRspackTest());
+
+  // TODO(Coly010): investigate this failure
+  xit('should support generating host and remote apps with the new name and root format', async () => {
+    const shell = uniq('shell');
+    const remote = uniq('remote');
+
+    runCLI(`generate @nx/react:host ${shell} --no-interactive --skipFormat`);
+    runCLI(
+      `generate @nx/react:remote ${remote} --host=${shell} --bundler=rspack --no-interactive --skipFormat`
+    );
+
+    const shellPort = readPort(shell);
+    const remotePort = readPort(remote);
+
+    // check files are generated without the layout directory ("apps/") and
+    // using the project name as the directory when no directory is provided
+    checkFilesExist(`${shell}/module-federation.config.ts`);
+    checkFilesExist(`${remote}/module-federation.config.ts`);
+
+    // check default generated host is built successfully
+    const buildOutputSwc = runCLI(`run ${shell}:build:development`);
+    expect(buildOutputSwc).toContain('Successfully ran target build');
+
+    const buildOutputTsNode = runCLI(`run ${shell}:build:development`, {
+      env: { NX_PREFER_TS_NODE: 'true' },
+    });
+    expect(buildOutputTsNode).toContain('Successfully ran target build');
+
+    // check serves devRemotes ok
+    const shellProcessSwc = await runCommandUntil(
+      `serve ${shell} --devRemotes=${remote} --verbose`,
+      (output) => {
+        return output.includes(
+          `All remotes started, server ready at http://localhost:${shellPort}`
+        );
+      }
+    );
+    await killProcessAndPorts(
+      shellProcessSwc.pid,
+      shellPort,
+      remotePort + 1,
+      remotePort
+    );
+
+    const shellProcessTsNode = await runCommandUntil(
+      `serve ${shell} --devRemotes=${remote} --verbose`,
+      (output) => {
+        return output.includes(
+          `All remotes started, server ready at http://localhost:${shellPort}`
+        );
+      },
+      {
+        env: { NX_PREFER_TS_NODE: 'true' },
+      }
+    );
+    await killProcessAndPorts(
+      shellProcessTsNode.pid,
+      shellPort,
+      remotePort + 1,
+      remotePort
+    );
+  }, 500_000);
+});
