+++ e2e/react/src/module-federation/core-rspack-setup.ts
@@ -1,0 +1,9 @@
+import { cleanupProject, newProject } from '@nx/e2e-utils';
+
+export function setupCoreRspackTest(): void {
+  newProject({ packages: ['@nx/react'] });
+}
+
+export function cleanupCoreRspackTest(): void {
+  cleanupProject();
+}
