+++ e2e/react/src/module-federation/core-webpack-setup.ts
@@ -1,0 +1,9 @@
+import { cleanupProject, newProject } from '@nx/e2e-utils';
+
+export function setupCoreWebpackTest(): void {
+  newProject({ packages: ['@nx/react', '@nx/webpack'] });
+}
+
+export function cleanupCoreWebpackTest(): void {
+  cleanupProject();
+}
