+++ e2e/workspace-create/src/create-nx-workspace-angular.test.ts
@@ -1,0 +1,97 @@
+import {
+  checkFilesDoNotExist,
+  checkFilesExist,
+  cleanupProject,
+  expectCodeIsFormatted,
+  getSelectedPackageManager,
+  runCreateWorkspace,
+  uniq,
+} from '@nx/e2e-utils';
+
+describe('create-nx-workspace --preset=angular', () => {
+  const packageManager = getSelectedPackageManager() || 'pnpm';
+
+  afterEach(() => cleanupProject());
+
+  it('should create a workspace with a single angular app at the root without routing', () => {
+    const wsName = uniq('angular');
+
+    runCreateWorkspace(wsName, {
+      preset: 'angular-standalone',
+      appName: wsName,
+      style: 'css',
+      packageManager,
+      standaloneApi: false,
+      routing: false,
+      e2eTestRunner: 'none',
+      bundler: 'webpack',
+      ssr: false,
+    });
+
+    checkFilesExist('package.json');
+    checkFilesExist('project.json');
+    checkFilesExist('src/app/app-module.ts');
+    checkFilesDoNotExist('src/app/app.routes.ts');
+    expectCodeIsFormatted();
+  });
+
+  it('should create a workspace with a single angular app at the root using standalone APIs', () => {
+    const wsName = uniq('angular');
+
+    runCreateWorkspace(wsName, {
+      preset: 'angular-standalone',
+      appName: wsName,
+      style: 'css',
+      packageManager,
+      standaloneApi: true,
+      routing: true,
+      e2eTestRunner: 'none',
+      bundler: 'webpack',
+      ssr: false,
+    });
+
+    checkFilesExist('package.json');
+    checkFilesExist('project.json');
+    checkFilesExist('src/app/app.routes.ts');
+    checkFilesDoNotExist('src/app/app-module.ts');
+    expectCodeIsFormatted();
+  });
+
+  it('should be able to create an angular workspace', () => {
+    const wsName = uniq('angular');
+    const appName = uniq('app');
+    runCreateWorkspace(wsName, {
+      preset: 'angular-monorepo',
+      style: 'css',
+      appName,
+      packageManager,
+      standaloneApi: false,
+      routing: true,
+      e2eTestRunner: 'none',
+      bundler: 'webpack',
+      ssr: false,
+    });
+    expectCodeIsFormatted();
+  });
+
+  it('should fail correctly when preset errors', () => {
+    // Using Angular Preset as the example here to test
+    // It will error when prefix is not valid
+    const wsName = uniq('angular-1-test');
+    const appName = uniq('app');
+    expect(() =>
+      runCreateWorkspace(wsName, {
+        preset: 'angular-monorepo',
+        style: 'css',
+        appName,
+        packageManager,
+        standaloneApi: false,
+        routing: false,
+        e2eTestRunner: 'none',
+        bundler: 'webpack',
+        ssr: false,
+        prefix: '1-one',
+      })
+    ).toThrow();
+  });
+});
