+++ e2e/workspace-create/src/create-nx-workspace-other-presets.test.ts
@@ -1,0 +1,147 @@
+import {
+  checkFilesDoNotExist,
+  checkFilesExist,
+  cleanupProject,
+  expectCodeIsFormatted,
+  expectNoAngularDevkit,
+  getSelectedPackageManager,
+  packageManagerLockFile,
+  runCreateWorkspace,
+  uniq,
+} from '@nx/e2e-utils';
+
+describe('create-nx-workspace --preset=other', () => {
+  const packageManager = getSelectedPackageManager() || 'pnpm';
+
+  afterEach(() => cleanupProject());
+
+  it('should reject workspace names starting with numbers', () => {
+    expect(() => {
+      runCreateWorkspace('4invalidname', {
+        preset: 'apps',
+        packageManager,
+      });
+    }).toThrow();
+  });
+
+  it('should be able to create an empty workspace built for apps', () => {
+    const wsName = uniq('apps');
+    runCreateWorkspace(wsName, {
+      preset: 'apps',
+      packageManager,
+    });
+
+    checkFilesExist('package.json', packageManagerLockFile[packageManager]);
+
+    expectNoAngularDevkit();
+  });
+
+  it('should be able to create an empty workspace with npm capabilities', () => {
+    const wsName = uniq('npm');
+    runCreateWorkspace(wsName, {
+      preset: 'npm',
+      packageManager,
+    });
+
+    expectNoAngularDevkit();
+    checkFilesDoNotExist('tsconfig.base.json');
+  });
+
+  it('should be able to create an empty workspace with ts/js capabilities', () => {
+    const wsName = uniq('ts');
+    runCreateWorkspace(wsName, {
+      preset: 'ts',
+      packageManager,
+    });
+
+    expectNoAngularDevkit();
+    expectCodeIsFormatted();
+  });
+
+  it('should be able to create an web-components workspace', () => {
+    const wsName = uniq('web-components');
+    const appName = uniq('app');
+    runCreateWorkspace(wsName, {
+      preset: 'web-components',
+      style: 'css',
+      appName,
+      packageManager,
+    });
+
+    expectNoAngularDevkit();
+    expectCodeIsFormatted();
+  });
+
+  it('should be able to create an express workspace', () => {
+    const wsName = uniq('express');
+    const appName = uniq('app');
+    runCreateWorkspace(wsName, {
+      preset: 'express',
+      docker: false,
+      appName,
+      packageManager,
+    });
+
+    expectNoAngularDevkit();
+    expectCodeIsFormatted();
+  });
+
+  it('should be able to create react-native workspace', () => {
+    const wsName = uniq('react-native');
+    const appName = uniq('app');
+    runCreateWorkspace(wsName, {
+      preset: 'react-native',
+      appName,
+      packageManager: 'npm',
+      e2eTestRunner: 'none',
+    });
+
+    expectNoAngularDevkit();
+    expectCodeIsFormatted();
+  });
+
+  it('should be able to create an expo workspace', () => {
+    const wsName = uniq('expo');
+    const appName = uniq('app');
+    runCreateWorkspace(wsName, {
+      preset: 'expo',
+      appName,
+      packageManager: 'npm',
+      e2eTestRunner: 'none',
+    });
+
+    expectNoAngularDevkit();
+    expectCodeIsFormatted();
+  });
+
+  it('should be able to create a workspace with a custom base branch and HEAD', () => {
+    const wsName = uniq('branch');
+    runCreateWorkspace(wsName, {
+      preset: 'apps',
+      base: 'main',
+      packageManager,
+    });
+  });
+
+  it('should be able to create a workspace with custom commit information', () => {
+    const wsName = uniq('branch');
+    runCreateWorkspace(wsName, {
+      preset: 'apps',
+      extraArgs:
+        '--commit.name="John Doe" --commit.email="myemail@test.com" --commit.message="Custom commit message!"',
+      packageManager,
+    });
+  });
+
+  it('should be able to create a nest workspace', () => {
+    const wsName = uniq('nest');
+    const appName = uniq('app');
+    runCreateWorkspace(wsName, {
+      preset: 'nest',
+      docker: false,
+      appName,
+      packageManager,
+    });
+    expectCodeIsFormatted();
+  });
+});
